runtimes/go: map context errors from the sqldb (#2129)
Currently all errors from the db are by defauled mapped to the
"Unavailable" encore error. This includes context errors.

For "canceled" in particular, I don't think unavailable is the correct
error code to return because it isn't indicative of an error with the
system. This is a pain for monitoring, where we want to excluded
specific codes like "canceled", but it does not cover DB errors.

So this change maps the context errors to equivalent encore error codes.

_This is technically a breaking change to the API - not sure if you
think the improvement is worth it_.