Add `parent_span_id` to protobuf def of SpanSummary for Traceview (#2188)
We need `parent_span_id` to build traceview from only a trace with a
specific span. This enables us to write `parent_span_id` and consume for
spans in the frontend where we want to view the spans to build the
spantree.

This adds
- `parent_span_id` to protobuf on trace2.proto
- adds column to database
- adds write to column of the database
- `emitCompleteSpanToListeners()`, WebSocket events will have NULL
parent_span_id (acceptable as it is not used by frontend)
- does not add any backfill for `Get`. Get uses events for populating
- does not add any backfill for `List`, see questions answered

**Questions answered:**
why do we not need to backFill `parent_span_id` on the spans returned by
List?
It is because list filters on `is_root` , is_root is always the root
span there the parent_id is "". Therefore we do not need to backfill the
spans returned in **List**.

TODO:
we need to investigate how that renders itself for the new frontend,
that requires `parent_span_id` to be set. Perhaps the implementation of
traces/ in **devdash** is different and uses the `is_root` property. See
[List](https://github.com/encoredev/encore/blob/main/cli/daemon/engine/trace2/sqlite/read.go#L51)
for filtering